{"schema":"decodex/commit/1","summary":"Modularize GraphRAG smoke runner","authority":"manual"}
