+++ scripts/graphrag_smoke/__init__.py
@@ -1,0 +1,1 @@
+"""GraphRAG Docker smoke modules."""
